moving rabbitmq to its own folder
