+++ services/web/server/src/simcore_service_webserver/rabbitmq/rabbitmq_settings.py
@@ -4,16 +4,15 @@
     - settings
 """
 
-from typing import Optional
 
 from aiohttp.web import Application
 from settings_library.rabbit import RabbitSettings
 
-from ._constants import APP_SETTINGS_KEY
+from .._constants import APP_SETTINGS_KEY
 
 
 def get_plugin_settings(app: Application) -> RabbitSettings:
-    settings: Optional[RabbitSettings] = app[APP_SETTINGS_KEY].WEBSERVER_RABBITMQ
+    settings: RabbitSettings | None = app[APP_SETTINGS_KEY].WEBSERVER_RABBITMQ
     assert settings, "setup_settings not called?"  # nosec
     assert isinstance(settings, RabbitSettings)  # nosec
     return settings
